split extension allow list plugin
