+++ packages/modular-scripts/src/esbuild-scripts/plugins/workerFactoryPlugin.ts
@@ -1,6 +1,7 @@
 import * as esbuild from 'esbuild';
 import * as path from 'path';
 import getModularRoot from '../../utils/getModularRoot';
+import createExtensionAllowlistPlugin from './extensionAllowList';
 
 // This plugin builds Web Workers on the fly and exports them to use like worker-loader for Webpack 4: https://v4.webpack.js.org/loaders/worker-loader/
 // The workers are not inlined, a new file is generated in the bundle. Only files *imported* with the *.worker pattern are matched.
@@ -113,29 +114,6 @@ function createPlugin(): esbuild.Plugin {
   };
 
   return plugin;
-}
-
-function createExtensionAllowlistPlugin(
-  allowedExtensions = ['.js', '.jsx', '.ts', '.tsx'],
-): esbuild.Plugin {
-  return {
-    name: 'worker-factory-plugin',
-    setup(build) {
-      // No lookbehind in Go regexp; need to look at all the files and do the check manually.
-      build.onResolve({ filter: /.*/ }, (args) => {
-        // Extract the extension; if not in the allow list, throw.
-        const extension = path.extname(args.path);
-        if (extension && !allowedExtensions.includes(extension)) {
-          throw new Error(
-            `Extension "${extension}" not allowed in web worker ${
-              args.importer
-            }. Permitted extensions are ${JSON.stringify(allowedExtensions)}`,
-          );
-        }
-        return undefined;
-      });
-    },
-  };
 }
 
 export default createPlugin;
